added some input validation

+++ persistence/src/main/java/com/epam/esm/dao/TagDao.java
@@ -18,5 +18,5 @@ public interface TagDao extends BaseDao<Tag> {
     Optional<Tag> findByName(String name);
 
     // TODO: 12/28/2021 add JD
-    Optional<Tag> findMostUsedTag(int id);
+    Optional<Tag> findMostUsedTag(long id);
 }
